requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioresponses==0.7.8
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.1
defusedxml==0.7.1
dnspython==2.7.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpretty==1.1.4
httpx==0.28.1
idna==3.10
ipwhois==1.3.0
lxml==5.4.0
multidict==6.4.3
packaging==25.0
propcache==0.3.1
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-Wappalyzer==0.3.1
python-whois==0.9.5
PyYAML==6.0.2
requests==2.32.3
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
